PM-3241: Move composition and options to a separate class for future reuse.
